refactor: provider이름 변경
